Factorize
changelog_begin
changelog_end
